+++ w09h01/test/pgdp/pingutrip/LegalCheck.java
@@ -12,7 +12,7 @@
  * THIS TEST CHECKS FOR INSTANCES OF KEYWORDS WITHIN YOUR   *
  * CODE. IN THE CASE OF FALSE POSITIVE UNCOMMENT LINE 37    *
  * AT YOUR OWN RISK                                         *
- *                                                          */                                                       
+ *                                                          */
 // ******************************************************* //
 
 package pgdp.pingutrip;
